Don't embed the manifest file, provide it externally.
The motivation behind this is to allow users to set uiAccess to true.

Setting uiAccess to true would allow JL to keep itself on top of pretty much everything (including an "Always on top" Task Manager) but in order for uiAccess to take effect, JL needs to be digitally signed and placed in a secure location (https://learn.microsoft.com/en-us/windows/security/threat-protection/security-policy-settings/user-account-control-only-elevate-uiaccess-applications-that-are-installed-in-secure-locations).

I am not willing to pay for the certificate but users have the option to use a self-signed certificate if they wish to use JL while uiAccess is set to true. Here's a good guide on how to do that: https://support.smartbear.com/testcomplete/docs/working-with/automating/via-com/configuring-manifests.html#using-extended-privilege-level---setting-the-uiaccess-attribute