Change the workaround used to prevent the text from being copied twice if it's copied from a WPF TextBox.
Earlier workaround was to check the time between two copies and discard the second result if the difference was less than 5 ms. This approach however was not very reliable because the double copy bug could occur after 5 ms if the copied text was long enough. And even more importantly, some scripts that copy the text to clipboard seem to do their job in an (at least seemingly) spammy manner and the aforementioned workaround prevents JL from getting the text from clipboard reliably when those scripts are used to copy the text to clipboard.

The new workaround is using a handcrafted copy function instead of relying WPF TextBox's built-in copy command.

Fixes #71.