PR Review Changes
Moved Icon to the end of the row so we can actually toggle visibility of the icon instead of uncollapse it
PR review changes like return types, array instead of List, etc.
Changed to Textblock instead of button
Made method to get first mining param instead of getting everything and then only getting the first.