Add "Click audio button hotkey" hotkey
Rename "Play audio" hotkey to "Play audio for the first reading"

Add "Click mining button hotkey" hotkey

Rename "Select the next/previous lookup result" hotkeys to "Select next/previous item". If the reading selection window for audio is open, they can be used to select the next/previous reading, if the spelling selection window for mining is open, they can be used to select the next/previous spelling, if neither of those windows are open, but a popup is open in mining mode, they can be used tol select the next/previous lookup result

Rename "Mine selected lookup result" hotkey to "Confirm selected item". It interacts with the selected item based on the active window. If the reading selection window for audio is open, it plays the selected reading. If the spelling selection window for mining is open, it mines the selected spelling. If neither of those windows are open, but a popup is open in mining mode, it mines the selected entry using its primary spelling.

Make "Add name", "Add word",  "Search", "Play audio for the first reading" hotkeys use the selected lookup result and if there's no selected lookup result, fallback to the first lookup result