Make "Focus on hover" option actually work
It turns out the SetForegroundWindow method almost always works when a debugger is attached to the current program and almost never works without it. The workaround is to attach JL to the foreground window's thread, steal focus from it, and then detach from its thread. While this is more intrusive than I’d prefer, it’s either this solution or removing this option entirely, as it won't work otherwise.