Don’t count user as idle before 1.5 seconds
For very short texts, such as a single kanji, the calculated idle timeout can become unrealistically short, sometimes even shorter than basic reaction and perception time, even for users with a high reading speed. To address this, a minimum idle threshold of 1.5 seconds is enforced. The value of this threshold is subject to change.