Treat IDE0005 as a suggestion to avoid build failures with PublishSingleFile
IDE0005 requires GenerateDocumentationFile to be true, but we don’t want those XML files in a release, so we set CopyDocumentationFileToOutputDirectory to false. This works when PublishSingleFile is not used, but causes the build to fail when it is. Until Roslyn resolves the issue, treat IDE0005 as a suggestion. See: https://github.com/dotnet/roslyn/issues/41640